Cleaned up various bits.
1.) Changed test_get_cell_temp() to get_cell_temp_fe6()
2.) Removed rdcfga's (get_cell_temp_fe6()
3.) Optimized get_cell_temp_fe6() delays by only doing adc on GPIO1 (was done an all GPIO and 2nd Ref previously)
4.) Cleaned up function parameters for get_cell_temp_fe6()
5.) Removed LTC6804_adax_fe6() function
6.) Removed unused and unnecessary variables

+++ alpaca-bms/alpaca-bms.cydsn/LTC68042.h
@@ -214,12 +214,10 @@ enum AuxPins {
     VREF2
 };
 
-uint16_t LTC6804_rdaux_fe6(enum AuxPins pin, uint16_t aux[3]);
+int8_t LTC6804_rdaux_fe6(enum AuxPins pin, uint16_t aux[3]);
 void LTC6804_adow(uint8_t pup);
 
 void LTC6804_adax();
-
-
 
 uint8_t LTC6804_rdcv(uint8_t reg, uint8_t total_ic, uint16_t cell_codes[][12]);
